validate DB connections and boot up process
Until now, every attempt to connect to the database was assumed to
work well, and every attempt to boot up the MUD was assumed to end up
successfully.

This commit changes that by making those two functions return a bool
value, which is also now validated by its callers.

This also works as groundwork for further validations tha might lead
to future enhancements, like #4.

+++ code/devextra.c
@@ -166,8 +166,10 @@ void clean_mud()
 	DESCRIPTOR_DATA *d;
 
 	DbConnection riftCore = nSQL.Settings.GetDbConnection("rift");
-	nSQL.StartSQLServer(riftCore.Host.c_str(),
-		riftCore.Db.c_str(), riftCore.User.c_str(), riftCore.Pwd.c_str());
+	if (!nSQL.StartSQLServer(riftCore.Host.c_str(),
+		riftCore.Db.c_str(), riftCore.User.c_str(), riftCore.Pwd.c_str()))
+		return RS.Bug("clean_mud: failed to establish a database connection.");
+
 	// 5184000 = one month
 
 	sprintf(buf, "DELETE FROM logins WHERE ctime + 5184000 < %ld", current_time);
